Lambda function triggered by Deeplens face detection
Lambda function that is triggered by S3 bucket 'deeplensfacedetected' upload. It calls Amazon Rekognition to compare the face Deeplens detected to the collection of indexes in the database of missing people. If a match is found, the information is stored in a DynamoDB database and a text message is sent to alert the user.